Refactor SessionMetaWorkspace and GroupSessionWorkspace
